[engine][core] Move enum convert from RenderParameters to VariableSet.
